Add NetBSD support. Currently, the `raw_exec` is the only usable driver.
